requirements.txt
streamlit
python-dotenv
openai
requests
googlemaps
supabase
beautifulsoup4
tqdm
